+++ llama_index/indices/empty/base.py
@@ -65,7 +65,7 @@ class EmptyIndex(BaseIndex[EmptyIndexStruct]):
             documents (List[BaseDocument]): A list of documents.
 
         Returns:
-            IndexList: The created list index.
+            IndexList: The created summary index.
         """
         del nodes  # Unused
         index_struct = EmptyIndexStruct()