+++ llama_index/indices/postprocessor/node_recency.py
@@ -5,10 +5,7 @@ from typing import List, Optional, Set
 import numpy as np
 import pandas as pd
 
-try:
-    from pydantic.v1 import Field
-except ImportError:
-    from pydantic import Field
+from llama_index.bridge.pydantic import Field
 
 from llama_index.indices.postprocessor.types import BaseNodePostprocessor
 from llama_index.indices.query.schema import QueryBundle