+++ llama_index/llms/azure_openai.py
@@ -1,9 +1,6 @@
 from typing import Any, Dict, Optional
 
-try:
-    from pydantic.v1 import Field
-except ImportError:
-    from pydantic import Field
+from llama_index.bridge.pydantic import Field
 
 from llama_index.callbacks import CallbackManager
 from llama_index.llms.openai import OpenAI