+++ llama_index/response_synthesizers/refine.py
@@ -1,11 +1,7 @@
 import logging
 from typing import Any, Generator, Optional, Sequence, cast, Type, Callable
 
-try:
-    from pydantic.v1 import BaseModel, Field
-except ImportError:
-    from pydantic import BaseModel, Field
-
+from llama_index.bridge.pydantic import BaseModel, Field, ValidationError
 from llama_index.indices.service_context import ServiceContext
 from llama_index.prompts.base import BasePromptTemplate, PromptTemplate
 from llama_index.llm_predictor.base import BaseLLMPredictor
@@ -173,12 +169,17 @@ class Refine(BaseSynthesizer):
         for cur_text_chunk in text_chunks:
             query_satisfied = False
             if response is None and not self._streaming:
-                structured_response = cast(
-                    StructuredRefineResponse, program(context_str=cur_text_chunk)
-                )
-                query_satisfied = structured_response.query_satisfied
-                if query_satisfied:
-                    response = structured_response.answer
+                try:
+                    structured_response = cast(
+                        StructuredRefineResponse, program(context_str=cur_text_chunk)
+                    )
+                    query_satisfied = structured_response.query_satisfied
+                    if query_satisfied:
+                        response = structured_response.answer
+                except ValidationError as e:
+                    logger.warning(
+                        f"Validation error on structured response: {e}", exc_info=True
+                    )
             elif response is None and self._streaming:
                 response = self._service_context.llm_predictor.stream(
                     text_qa_template,
@@ -227,12 +228,17 @@ class Refine(BaseSynthesizer):
         for cur_text_chunk in text_chunks:
             query_satisfied = False
             if not self._streaming:
-                structured_response = cast(
-                    StructuredRefineResponse, program(context_msg=cur_text_chunk)
-                )
-                query_satisfied = structured_response.query_satisfied
-                if query_satisfied:
-                    response = structured_response.answer
+                try:
+                    structured_response = cast(
+                        StructuredRefineResponse, program(context_msg=cur_text_chunk)
+                    )
+                    query_satisfied = structured_response.query_satisfied
+                    if query_satisfied:
+                        response = structured_response.answer
+                except ValidationError as e:
+                    logger.warning(
+                        f"Validation error on structured response: {e}", exc_info=True
+                    )
             else:
                 response = self._service_context.llm_predictor.stream(
                     refine_template,
@@ -303,13 +309,20 @@ class Refine(BaseSynthesizer):
         for cur_text_chunk in text_chunks:
             query_satisfied = False
             if not self._streaming:
-                structured_response = await program.acall(context_msg=cur_text_chunk)
-                structured_response = cast(
-                    StructuredRefineResponse, structured_response
-                )
-                query_satisfied = structured_response.query_satisfied
-                if query_satisfied:
-                    response = structured_response.answer
+                try:
+                    structured_response = await program.acall(
+                        context_msg=cur_text_chunk
+                    )
+                    structured_response = cast(
+                        StructuredRefineResponse, structured_response
+                    )
+                    query_satisfied = structured_response.query_satisfied
+                    if query_satisfied:
+                        response = structured_response.answer
+                except ValidationError as e:
+                    logger.warning(
+                        f"Validation error on structured response: {e}", exc_info=True
+                    )
             else:
                 raise ValueError("Streaming not supported for async")
 
@@ -337,13 +350,20 @@ class Refine(BaseSynthesizer):
         # TODO: consolidate with loop in get_response_default
         for cur_text_chunk in text_chunks:
             if response is None and not self._streaming:
-                structured_response = await program.acall(context_str=cur_text_chunk)
-                structured_response = cast(
-                    StructuredRefineResponse, structured_response
-                )
-                query_satisfied = structured_response.query_satisfied
-                if query_satisfied:
-                    response = structured_response.answer
+                try:
+                    structured_response = await program.acall(
+                        context_str=cur_text_chunk
+                    )
+                    structured_response = cast(
+                        StructuredRefineResponse, structured_response
+                    )
+                    query_satisfied = structured_response.query_satisfied
+                    if query_satisfied:
+                        response = structured_response.answer
+                except ValidationError as e:
+                    logger.warning(
+                        f"Validation error on structured response: {e}", exc_info=True
+                    )
             elif response is None and self._streaming:
                 raise ValueError("Streaming not supported for async")
             else: