+++ llama_index/vector_stores/__init__.py
@@ -29,7 +29,10 @@ from llama_index.vector_stores.docarray import (
     DocArrayHnswVectorStore,
     DocArrayInMemoryVectorStore,
 )
+from llama_index.vector_stores.cogsearch import CognitiveSearchVectorStore
 from llama_index.vector_stores.awadb import AwaDBVectorStore
+from llama_index.vector_stores.neo4jvector import Neo4jVectorStore
+from llama_index.vector_stores.epsilla import EpsillaVectorStore
 
 __all__ = [
     "SimpleVectorStore",
@@ -57,4 +60,7 @@ __all__ = [
     "ZepVectorStore",
     "AwaDBVectorStore",
     "BagelVectorStore",
+    "Neo4jVectorStore",
+    "CognitiveSearchVectorStore",
+    "EpsillaVectorStore",
 ]